Cache preprocessed JMeter Reports (JENKINS-9031)
Instead of parsing all available JTL results upon each request,
we now save the parsed report as a serialized Java object for later
reuse. The file will be next to the JTL file with a '.serialized'
suffix. Additionally the report also gets added to a in-memory cache
with a maximum capacity of 100 reports. This considerably speeds up the
performance trend display.

Signed-off-by: Michel Marti <mma@objectxp.com>