Wrk Summarizer Parser
`wrk` is a utility for load testing a web-service over http/https.
`wrk` is similar in functionality to Apache's `ab` but differs in
that it uses event-notification systems and threads to create a more
realistic load from a single machine (thus eliminating/delaying the
need for master-slave configurations).

A parse was added to parse the output from `wrk` (generally sent to
STDOUT). This output is a summarized format and the parser generates
PerformanceReport(s) similar to how the existing JUnitSummarizerParser
does.

+ Added Summarizer Parser for `wrk`
+ Added test suite for new WrkSummarizerParser class
+ Added wrk output-files for testing (added to src/test/resources)