Compatible with flexible publisher
The original parser lookup in ifSummarizerParsedUsed is not comptaible with flexible publisher, because the performance publisher is not directly in buildAction.getBuild().getProject().getPublishersList(), but wrapped inside publisher list inside flexible publisher. It will result in NPE because buildAction.getBuild().getProject().getPublishersList().get(PerformancePublisher.class) return null.

We can just simply look up with buildAction.getParserByDisplayName directly instead of looping the project publisher list to check for JmeterSummarizer and Iago parser. The parser list is already there in buildAction. This way the code is simpler, and we can avoid complicated lookup when this publisher is under flexible publisher.