[FIXES JENKINS-28787]: Fixes number format exception
When JMeter parser is used and a csv file contains header line, a number format exception is thrown.
Fix is to check if the first line is a header and skip the line if it is a header.
Note: Noticed that if there was a test for the constructor with glob only, this issue
would have been caught in testing.

Test added and refactored for cleanup.