Fix syntax and snippet generator in Pipeline
The previous approach was...odd. The result was that the snippet
generator would fail badly, and there were a lot of "required"
parameters that were not, in fact, required. So I reworked things to
make the non-required parameters not on the @DataBoundConstructors,
added @DataBoundSetter methods where appropriate, added the additional
"bzt" and "perfReport" symbols for compatibility purposes, added real
tests for the steps, and bumped a lot of dependencies to modern
versions, etc.

This does *not* fix the durability problem - PerformancePublisher
isn't a DurableTask, and so if it's running while the Jenkins master
restarts or otherwise gets disconnected from the agent, the build will
fail. With a DurableTask, such as "sh" or "bat", the task will keep
running even if the master gets disconnected, and the build will pick
back up upon reconnecting.