Test with Java 21
Java 21 was released Sep 19, 2023. We want to announce full support for
Java 21 in October and would like the most used plugins to be compiled
and tested with Java 21.

The acceptance test harness and plugin bill of materials tests are already
passing with Java 21. This is a further step to improve plugin readiness
for use with Java 21 and for development with Java 21.

The change intentionally tests only one Java configuration, Java 21
because the risk of a regression that only affects Java 11 is low. We
generate Java 11 byte code with the Java 21 build, so we're already
testing Java 11 byte code.