Test with Java 25
Java 25 released September 16, 2025.  The Jenkins project wants to support
Java 25 soon.  Compile and test on ci.jenkins.io with Java 25 and Java 21.

Intentionally continues to generate Java 17 byte code as configured by
the plugin parent pom.

Does not compile or test with Java 21 any longer because we have found no
issues in the past that were specific to the Java 21 compiler.

Testing done:

* Rely on ci.jenkins.io because my local Python environment does not
  match the needs of the plugin